requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0
python-dotenv==1.0.1
huggingface-hub==0.23.4
